-- imported changes thelib and crtmpserver from commercial version (evostream)
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@734 01689507-d990-df11-9d7d-0017a411d416

+++ sources/thelib/src/protocols/liveflv/innetliveflvstream.cpp
@@ -108,10 +108,10 @@ void InNetLiveFLVStream::GetStats(Variant &info, uint32_t namespaceId) {
 	BaseInNetStream::GetStats(info, namespaceId);
 	info["audio"]["bytesCount"] = _audioBytesCount;
 	info["audio"]["packetsCount"] = _audioPacketsCount;
-	info["audio"]["droppedPacketsCount"] = 0;
+	info["audio"]["droppedPacketsCount"] = (uint32_t) 0;
 	info["video"]["bytesCount"] = _videoBytesCount;
 	info["video"]["packetsCount"] = _videoPacketsCount;
-	info["video"]["droppedPacketsCount"] = 0;
+	info["video"]["droppedPacketsCount"] = (uint32_t) 0;
 }
 
 void InNetLiveFLVStream::SignalOutStreamAttached(BaseOutStream *pOutStream) {
